Simple_Quiz_app
I have used Android Studio Software for designing applications. I have created an app using Java in Android Studio. I have uploaded both Java and Xml files. This is a Quiz application. It is a simple quiz app. Actually, I have designed simple quiz app but for some more updating in app, I have taken some help from the tutorials for final Score of the quiz. I have used another class for questions and options. Instead of finding questions in Main Activity file for changing the questions, I have used question.java file for changing the questions. It was easy to everyone for changing the questions for different quizzes. The final Output of this app gives an Alert dialog box, that shows Score of the quiz and restart the quiz option also.