Calculator_app
I have designed this app using Android Studio software. This application is Basic Arithmetic Calculator. It is helpful to calculate the basic calculations like addition, subtraction, multiplication and division. The benefit of this calculator is that it can give the output in decimal values and also it can take the input in decimal values.